Fix ModelYolo possible loss of data warning
Cherry pick
6b126df45492c2316664163af6dcd921b2b58383

+++ demos/common/cpp/models/src/detection_model_yolo.cpp
@@ -315,7 +315,7 @@ void ModelYolo::parseYOLOOutput(const std::string& output_name,
                 obj.width = clamp(width, 0.f, (float)original_im_w - obj.x);
                 obj.height = clamp(height, 0.f, (float)original_im_h - obj.y);
 
-                for (int j = 0; j < region.classes; ++j) {
+                for (size_t j = 0; j < region.classes; ++j) {
                     int class_index = calculateEntryIndex(entriesNum, region.coords, region.classes + isObjConf, n * entriesNum + i, region.coords + isObjConf + j);
                     float prob = scale * postprocessRawData(outData[class_index]);
 
@@ -332,7 +332,7 @@ void ModelYolo::parseYOLOOutput(const std::string& output_name,
     }
 }
 
-int ModelYolo::calculateEntryIndex(int totalCells, int lcoords, int lclasses, int location, int entry) {
+int ModelYolo::calculateEntryIndex(int totalCells, int lcoords, size_t lclasses, int location, int entry) {
     int n = location / totalCells;
     int loc = location % totalCells;
     return (n * (lcoords + lclasses) + entry) * totalCells + loc;
@@ -377,7 +377,7 @@ ModelYolo::Region::Region(const std::shared_ptr<ov::op::v0::RegionYolo>& regionY
     }
 }
 
-ModelYolo::Region::Region(int classes, int coords, const std::vector<float>& anchors, const std::vector<int64_t>& masks, int outputWidth, int outputHeight) :
+ModelYolo::Region::Region(size_t classes, int coords, const std::vector<float>& anchors, const std::vector<int64_t>& masks, size_t outputWidth, size_t outputHeight) :
     classes(classes), coords(coords),
     outputWidth(outputWidth), outputHeight(outputHeight) {
     num = masks.size();
